+++ src/telegram/chat.h
@@ -14,7 +14,7 @@ struct Chat
 {
     typedef QSharedPointer<Chat> Ptr;
 
-    int32_t                 m_id = 0;
+    qint64                  m_id = 0;
     QString                 m_type;
     QString                 m_title;
     QString                 m_username;
@@ -27,12 +27,12 @@ struct Chat
     QString                 m_invite_link;
     QSharedPointer<Message> m_pinned_message;
     ChatPermissions::Ptr    m_permissions;
-    int32_t                 m_slow_mode_delay          = 0;
-    int32_t                 m_message_auto_delete_time = 0;
+    qint64                  m_slow_mode_delay          = 0;
+    qint64                  m_message_auto_delete_time = 0;
     bool                    m_has_protected_content    = false;
     QString                 m_sticker_set_name;
     bool                    m_can_set_sticker_set = false;
-    int32_t                 m_linked_chat_id      = 0;
+    qint64                  m_linked_chat_id      = 0;
     ChatLocation::Ptr       m_location;
 };
