+++ src/telegram/location.h
@@ -9,12 +9,12 @@ struct Location
 {
     typedef QSharedPointer<Location> Ptr;
 
-    float   m_longitude              = 0;
-    float   m_latitude               = 0;
-    float   m_horizontal_accuracy    = 0;
-    int32_t m_live_period            = 0;
-    int32_t m_heading                = 0;
-    int32_t m_proximity_alert_radius = 0;
+    double m_longitude              = 0;
+    double m_latitude               = 0;
+    double m_horizontal_accuracy    = 0;
+    qint64 m_live_period            = 0;
+    qint64 m_heading                = 0;
+    qint64 m_proximity_alert_radius = 0;
 };
 
 void readValue(Location::Ptr& value, const QJsonObject& json, const QString& valueName);
