+++ src/telegram/maskposition.h
@@ -10,9 +10,9 @@ struct MaskPosition
     typedef QSharedPointer<MaskPosition> Ptr;
 
     QString m_point;
-    float   m_x_shift = 0;
-    float   m_y_shift = 0;
-    float   m_scale   = 0;
+    double   m_x_shift = 0;
+    double   m_y_shift = 0;
+    double   m_scale   = 0;
 };
 
 void readValue(MaskPosition::Ptr& value, const QJsonObject& json, const QString& valueName);
