+++ src/telegram/successfulpayment.h
@@ -11,7 +11,7 @@ struct SuccessfulPayment
     typedef QSharedPointer<SuccessfulPayment> Ptr;
 
     QString        m_currency;
-    int32_t        m_total_amount = 0;
+    qint64         m_total_amount = 0;
     QString        m_invoice_payload;
     QString        m_shipping_option_id;
     OrderInfo::Ptr m_order_info;
