+++ src/telegram/videonote.h
@@ -12,10 +12,10 @@ struct VideoNote
 
     QString        m_file_id;
     QString        m_file_unique_id;
-    int32_t        m_length   = 0;
-    int32_t        m_duration = 0;
+    qint64         m_length   = 0;
+    qint64         m_duration = 0;
     PhotoSize::Ptr m_thumb;
-    int32_t        m_file_size = 0;
+    qint64         m_file_size = 0;
 };
 
 void readValue(VideoNote::Ptr& value, const QJsonObject& json, const QString& valueName);
