+++ src/telegram/voice.h
@@ -11,9 +11,9 @@ struct Voice
 
     QString m_file_id;
     QString m_file_unique_id;
-    int32_t m_duration = 0;
+    qint64  m_duration = 0;
     QString m_mime_type;
-    int32_t m_file_size = 0;
+    qint64  m_file_size = 0;
 };
 
 void readValue(Voice::Ptr& value, const QJsonObject& json, const QString& valueName);
